osiris/requirements.txt
numpy==1.18.4
pandas==1.0.3
plotly==4.7.1
python-dateutil==2.8.1
pytz==2020.1
retrying==1.3.3
six==1.15.0
